implement save/restore, and make image grid fancier
