+++ scripts/mnist_jem.py
@@ -8,9 +8,13 @@ Implements Stochastic gradient Langevin dynamics for energy-based models,
 as per https://arxiv.org/pdf/1912.03263.pdf
 """
 
+# TODO -- test EpochIndependentReplayBuffer
 # TODO - periodically, rank all of the MCMC samples by their probability, and the highest-probability samples
-
-import argparse
+# TODO -- implement checkpointing & automatic reversion to a saved model if some loss metric increases too much
+# TODO -- implement 3-way split for MNIST
+# TODO -- double batch size each epoch, up to some maximum -- e.g., from 8 to 128
+# TODO -- change how the grid view works so that you're always getting 36 images, regardless of batch size
+# TODO -- implement label replay buffer and label SGLD
 import datetime
 import pathlib
 
@@ -28,248 +32,7 @@ from torchvision.utils import make_grid
 from sgld_nrg.networks import Resnet, SimpleNet, ToyNet
 from sgld_nrg.sgld import IndependentReplayBuffer, SgldLogitEnergy
 from sgld_nrg.transform import AddGaussianNoise
-
-
-def nonnegative_int(value):
-    ivalue = int(value)
-    if ivalue < 0:
-        raise argparse.ArgumentTypeError(
-            f"provided value {value} is not a positive int"
-        )
-    return ivalue
-
-
-def positive_int(value):
-    ivalue = int(value)
-    if ivalue <= 0:
-        raise argparse.ArgumentTypeError(
-            f"provided value {value} is not a positive int"
-        )
-    return ivalue
-
-
-def positive_float(value):
-    fvalue = float(value)
-    if fvalue <= 0.0:
-        raise argparse.ArgumentTypeError(
-            f"provided value {value} is not a positive float"
-        )
-    return fvalue
-
-
-def nonnegative_float(value):
-    fvalue = float(value)
-    if fvalue < 0.0:
-        raise argparse.ArgumentTypeError(f"provided value {value} is a negative float")
-    return fvalue
-
-
-def parse_args():
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "-s",
-        "--save",
-        required=True,
-        type=pathlib.Path,
-        help="must provide a name for checkpoints; if the checkpoint already exists, the model is loaded",
-    )
-    parser.add_argument(
-        "--pretrain_energy_penalty",
-        default=0.0,
-        type=nonnegative_float,
-        help="how much to penalize the pre-training model using the energy; set to 0.0 (default) to disable",
-    )
-    parser.add_argument(
-        "--num_workers",
-        default=2,
-        type=nonnegative_int,
-        help="number of processes to spawn for DataLoader",
-    )
-    parser.add_argument(
-        "--translate",
-        default=0,
-        type=nonnegative_int,
-        help="randomly translates the image but up to this many pixels; set to 0 to disable",
-    )
-    parser.add_argument(
-        "--input_noise",
-        default=0.03,
-        type=positive_float,
-        help="the magnitude of 0-mean Gaussian noise applied to the MNIST images in each minibatch; set to 0 to disable",
-    )
-    parser.add_argument(
-        "--network",
-        default="simple",
-        choices=["simple", "residual", "toy"],
-        help="which neural network architecture to use for training; see networks.py",
-    )
-    parser.add_argument(
-        "--sgld_lr",
-        default=1.0,
-        type=positive_float,
-        help="the size of the SGLD gradient step",
-    )
-    parser.add_argument(
-        "--sgld_noise",
-        default=1e-2,
-        type=positive_float,
-        help="the standard deviation of the gaussian noise in SGLD",
-    )
-    parser.add_argument(
-        "-F",
-        "--fashion",
-        action="store_true",
-        help="set this flag to train with FashionMNIST instead of the MNIST digits",
-    )
-    parser.add_argument(
-        "-r",
-        "--lr",
-        default=1e-5,
-        type=float,
-        help="the learning rate for the optimizer",
-    )
-    parser.add_argument(
-        "--replay_buff",
-        default=10000,
-        type=positive_int,
-        help="How many chains to store in the SGLD MCMC buffer",
-    )
-    parser.add_argument(
-        "--prob_reinit",
-        default=0.05,
-        type=positive_float,
-        help="The probability of reinitializing an SGLD MCMC chain",
-    )
-    parser.add_argument(
-        "--sgld_steps",
-        default=15,
-        type=positive_int,
-        help="how many SGLD steps to take at each iteration",
-    )
-    parser.add_argument(
-        "-b", "--batch_size", default=8 * 8, type=positive_int, help="mini-batch size"
-    )
-    parser.add_argument(
-        "-e",
-        "--n_epochs",
-        default=10,
-        type=nonnegative_int,
-        help="number of epochs of SGLD training",
-    )
-    parser.add_argument(
-        "-p",
-        "--pre_epochs",
-        default=0,
-        type=nonnegative_int,
-        help="how many epochs to train the network as an ordinary classifier before using SGLD",
-    )
-    return parser.parse_args()
-
-
-def get_accuracy(y_hat, y):
-    _, predicted = torch.max(y_hat.data, 1)
-    correct = (predicted == y).sum().item()
-    return correct / y.size(0)
-
-
-def estimate_time_remaining(elapsed_seconds, batch_offset, num_batch):
-    multiplier = (1.0 - (batch_offset + 1) / num_batch) / (batch_offset + 1) * num_batch
-    return elapsed_seconds / 60.0 / 60.0 * multiplier
-
-
-def clf_train_one_epoch(writer):
-    pre_train_buff_size = max(1, int(512 / user_args.batch_size + 0.5))
-    pre_train_buff_xe = np.zeros(pre_train_buff_size)
-    pre_train_buff_acc = np.zeros(pre_train_buff_size)
-    pre_train_buff_pnrg = np.zeros(pre_train_buff_size)
-    pre_val_buff = np.zeros(len(test))
-    pre_val_buff_acc = np.zeros(len(test))
-    pre_val_buff_pnrg = np.zeros(len(test))
-    pre_train_counter = 0
-    pointer = 0
-    xe_loss = 0.0
-    dir = user_args.save.absolute().parent
-    stem = user_args.save.absolute().stem
-    print(stem)
-    for pre_train_epoch in range(user_args.pre_epochs):
-        print(f"Pre-train epoch {pre_train_epoch} of {user_args.pre_epochs}")
-        start_time = datetime.datetime.now()
-        for i, (x_train, y_train) in enumerate(train):
-            my_net.train()
-            pre_train_counter += x_train.size(0)
-            y_logit = my_net(x_train)
-            xe_loss = xe_loss_fn(y_logit, y_train)
-            pnrg = my_net.logsumexp_logits(x_train).mean()
-            total_loss = xe_loss + user_args.pretrain_energy_penalty * pnrg
-            total_loss.backward()
-            pre_train_optim.step()
-            pre_train_optim.zero_grad()
-            pre_train_buff_acc[pointer % pre_train_buff_size] = get_accuracy(
-                y_logit, y_train
-            )
-            pre_train_buff_xe[pointer % pre_train_buff_size] = xe_loss.item()
-            pre_train_buff_pnrg[pointer % pre_train_buff_size] = pnrg
-            pointer += 1
-            if i % pre_train_buff_size == 0 and i > 0:
-                elapsed = (datetime.datetime.now() - start_time) / datetime.timedelta(
-                    seconds=1
-                )
-                epoch_remaining = estimate_time_remaining(
-                    elapsed_seconds=elapsed, batch_offset=i, num_batch=len(train)
-                )
-                print(
-                    f"\tPre-train batch {i} of {len(train)} - est. {epoch_remaining:.3f} hours remaining"
-                )
-                writer.add_scalar(
-                    "pre/accuracy/train", pre_train_buff_acc.mean(), pre_train_counter
-                )
-                writer.add_scalar(
-                    "pre/xe_loss/train", pre_train_buff_xe.mean(), pre_train_counter
-                )
-                writer.add_scalar(
-                    "combined/seconds_per_instance",
-                    elapsed / ((i + 1) * user_args.batch_size),
-                    pre_train_counter,
-                )
-                writer.add_scalar(
-                    "combined/accuracy", pre_train_buff_acc.mean(), pre_train_counter
-                )
-                writer.add_scalar(
-                    "combined/xe", pre_train_buff_xe.mean(), pre_train_counter
-                )
-                writer.add_scalar(
-                    "combined/+nrg", pre_train_buff_pnrg.mean(), pre_train_counter
-                )
-                for name, weight in my_net.named_parameters():
-                    writer.add_histogram(f"param/{name}", weight, pre_train_counter)
-                    writer.add_histogram(f"grad/{name}", weight.grad, pre_train_counter)
-
-        for j, (x_val, y_val) in enumerate(test):
-            my_net.eval()
-            y_logit = my_net(x_val)
-            xe_loss = xe_loss_fn(y_logit, y_val)
-            pre_val_buff[j] = xe_loss.item()
-            pre_val_buff_acc[j] = get_accuracy(y_logit, y_val)
-            pre_val_buff_pnrg[j] = my_net.logsumexp_logits(x_val).mean()
-        writer.add_scalar(
-            "pre/accuracy/val", pre_val_buff_acc.mean(), pre_train_counter
-        )
-        writer.add_scalar("pre/xe_loss/val", pre_val_buff.mean(), pre_train_counter)
-        writer.add_scalar("pre/+nrg/val", pre_val_buff_pnrg.mean(), pre_train_counter)
-        fname = f"{stem}-pre-epoch-{pre_train_epoch}.pth"
-        pre_dest = dir.joinpath(fname)
-        print(f"\tSaving epoch {pre_train_epoch} to {pre_dest}")
-        torch.save(
-            {
-                "epoch": pre_train_epoch,
-                "model_state_dict": my_net.state_dict(),
-                "optimizer_state_dict": pre_train_optim.state_dict(),
-                "loss": xe_loss,
-            },
-            str(pre_dest),
-        )
-    return pre_train_counter
-
+from sgld_nrg.utils import estimate_time_remaining, get_accuracy, parse_args
 
 if __name__ == "__main__":
     user_args = parse_args()
@@ -327,7 +90,7 @@ if __name__ == "__main__":
         )
     train = DataLoader(
         train,
-        batch_size=user_args.batch_size,
+        batch_size=user_args.batch_size_min,
         shuffle=True,
         drop_last=True,
         num_workers=user_args.num_workers,
@@ -342,7 +105,7 @@ if __name__ == "__main__":
 
     if user_args.network == "simple":
         my_net = SimpleNet()
-    elif user_args.network == "residual":
+    elif user_args.network == "resnet":
         my_net = Resnet()
     elif user_args.network == "toy":
         my_net = ToyNet()
@@ -355,13 +118,6 @@ if __name__ == "__main__":
     torch_summary(my_net, (1, 28, 28))
     param_ct = sum(p.numel() for p in my_net.parameters() if p.requires_grad)
     print(f"The network has {param_ct:,} parameters.")
-    if user_args.save.exists() and user_args.save.absolute().is_file():
-        print(f"Loading saved model from {user_args.save}...")
-        checkpoint = torch.load(user_args.save)
-        my_net.load_state_dict(checkpoint["model_state_dict"])
-        main_optim.load_state_dict(checkpoint["optimizer_state_dict"])
-
-
     my_buffer = IndependentReplayBuffer(
         data_shape=(1, 28, 28),
         data_range=(-1.0, 1.0),
@@ -375,19 +131,29 @@ if __name__ == "__main__":
         noise=user_args.sgld_noise,
         sgld_step=user_args.sgld_steps,
     )
+    if user_args.save.exists() and user_args.save.absolute().is_file():
+        print(f"Loading saved model from {user_args.save}...")
+        checkpoint = torch.load(user_args.save)
+        my_net.load_state_dict(checkpoint["model_state_dict"])
+        main_optim.load_state_dict(checkpoint["optimizer_state_dict"])
+        my_sgld.replay.buffer = checkpoint["MCMC_samples"]
+    else:
+        print(
+            f"No saved model {user_args.save} found; training the model from its random initialization..."
+        )
+
     xe_loss_fn = nn.CrossEntropyLoss(reduction="mean")
 
     writer = SummaryWriter()
-    pre_train_optim = Adam(my_net.parameters(), user_args.lr)
-    pre_train_counter = clf_train_one_epoch(writer=writer)
 
-    sgld_train_buff = max(1, int(512 / user_args.batch_size + 0.5))
+    sgld_train_buff = max(4, int(512 / user_args.batch_size_min + 0.5))
     sgld_train_total_buff = np.zeros(sgld_train_buff)
     sgld_train_xe_buff = np.zeros(sgld_train_buff)
     sgld_train_nrg_buff = np.zeros(sgld_train_buff)
     sgld_train_acc_buff = np.zeros(sgld_train_buff)
     sgld_train_pnrg_buff = np.zeros(sgld_train_buff)
     sgld_train_nnrg_buff = np.zeros(sgld_train_buff)
+    sgld_sample_buff = torch.zeros(36, 1, 28, 28)
     jem_counter = 0
     total_loss = 0.0
     for epoch_num in range(user_args.n_epochs):
@@ -400,7 +166,7 @@ if __name__ == "__main__":
             y_logit = my_net(x_train)
             xe_loss = xe_loss_fn(y_logit, y_train)
             x_nrg = my_net.logsumexp_logits(x_train).mean()
-            x_hat = my_sgld(user_args.batch_size)
+            x_hat = my_sgld(user_args.batch_size_min)
             x_hat_nrg = my_net.logsumexp_logits(x_hat).mean()
             total_loss = xe_loss + x_nrg - x_hat_nrg
             total_loss.backward()
@@ -413,7 +179,12 @@ if __name__ == "__main__":
             sgld_train_pnrg_buff[i % sgld_train_buff] = x_nrg.item()
             sgld_train_nnrg_buff[i % sgld_train_buff] = x_hat_nrg.item()
             sgld_train_acc_buff[i % sgld_train_buff] = train_acc
-
+            if x_hat.size(0) < sgld_sample_buff.size(0):
+                into = torch.arange(0, x_hat.size(0)) + jem_counter
+                into %= sgld_sample_buff.size(0)
+                sgld_sample_buff[into, ...] = x_hat
+            else:
+                sgld_sample_buff = x_hat[: sgld_sample_buff.size(0), ...]
             if i % sgld_train_buff == 0:
                 elapsed_s = (datetime.datetime.now() - start_time) / datetime.timedelta(
                     seconds=1
@@ -425,28 +196,29 @@ if __name__ == "__main__":
                     f"\tBatch {i} of {len(train)}; estimate {remaining_h:.2f} hours remaining in this epoch"
                 )
                 x_hat_grid = make_grid(
-                    -1.0 * (x_hat * normalize_scale + normalize_center),
-                    nrow=min(10, int(np.sqrt(user_args.batch_size) + 0.5)),
-                )  # reverse the scaling applied earlier for display purposes
+                    -1.0
+                    * (
+                        sgld_sample_buff * normalize_scale + normalize_center
+                    ),  # reverse the scaling applied earlier for display purposes
+                    nrow=int(np.sqrt(sgld_sample_buff.size(0)) + 0.5),
+                )
                 writer.add_image("JEM/x_hat", x_hat_grid, jem_counter)
                 writer.add_scalar("JEM/total", total_loss.item(), jem_counter)
                 writer.add_scalar("JEM/xe", sgld_train_xe_buff.mean(), jem_counter)
-                writer.add_scalar("JEM/\u0394nrg", xe_loss.item(), jem_counter)
+                writer.add_scalar(
+                    "JEM/\u0394nrg", (x_nrg - x_hat_nrg).item(), jem_counter
+                )
                 writer.add_scalar("JEM/+nrg", x_nrg.item(), jem_counter)
                 writer.add_scalar("JEM/-nrg", x_hat_nrg.item(), jem_counter)
                 writer.add_scalar("JEM/accuracy", train_acc, jem_counter)
                 writer.add_scalar(
                     "combined/seconds_per_instance",
-                    elapsed_s / ((i + 1) * user_args.batch_size),
-                    jem_counter + pre_train_counter,
+                    elapsed_s / ((i + 1) * user_args.batch_size_min),
+                    jem_counter,
                 )
                 for name, weight in my_net.named_parameters():
-                    writer.add_histogram(
-                        f"param/{name}", weight, jem_counter + pre_train_counter
-                    )
-                    writer.add_histogram(
-                        f"grad/{name}", weight.grad, jem_counter + pre_train_counter
-                    )
+                    writer.add_histogram(f"param/{name}", weight, jem_counter)
+                    writer.add_histogram(f"grad/{name}", weight.grad, jem_counter)
 
             if i > 0 and i % sgld_train_buff == 0:
                 writer.add_scalar(
@@ -456,32 +228,33 @@ if __name__ == "__main__":
                 writer.add_scalar(
                     "JEM/\u0394nrg", sgld_train_nrg_buff.mean(), jem_counter
                 )
-                writer.add_scalar(
-                    "combined/+nrg", x_nrg.item(), jem_counter + pre_train_counter
-                )
+                writer.add_scalar("combined/+nrg", x_nrg.item(), jem_counter)
                 writer.add_scalar(
                     "combined/xe",
                     sgld_train_xe_buff.mean(),
-                    jem_counter + pre_train_counter,
+                    jem_counter,
                 )
-                writer.add_scalar(
-                    "combined/accuracy", train_acc, jem_counter + pre_train_counter
-                )
+                writer.add_scalar("combined/accuracy", train_acc, jem_counter)
                 writer.add_scalar("JEM/+nrg", sgld_train_pnrg_buff.mean(), jem_counter)
                 writer.add_scalar("JEM/-nrg", sgld_train_nnrg_buff.mean(), jem_counter)
                 writer.add_scalar(
                     "JEM/accuracy", sgld_train_acc_buff.mean(), jem_counter
                 )
+        for j, (x_test, y_test) in enumerate(test):
+            # TODO
+            break
         dir = user_args.save.absolute().parent
         stem = user_args.save.absolute().stem
         fname = f"{stem}-epoch-{epoch_num}.pth"
         dest = dir.joinpath(fname)
+        print(f"Saving model checkpoint to {dest}...")
         torch.save(
             {
                 "epoch": epoch_num,
                 "model_state_dict": my_net.state_dict(),
                 "optimizer_state_dict": main_optim.state_dict(),
                 "loss": total_loss,
+                "MCMC_samples": my_sgld.replay.buffer,
             },
             str(dest),
         )
